+++ app/assets/javascripts/application_v2.js
@@ -5,6 +5,7 @@
 //= require jquery_ujs
 
 //= require turbolinks
+//= require turbolinks_compatibility
 //= require react
 //= require react_ujs
 //= require components